change more skeletons
